Missing static files
There were some missing land card imgs for some reason. 
Also, moved prospecting canvas stuff to its own class, handle window resizing / drawing / scaling
