+++ src/vue/land.vue
@@ -4,14 +4,14 @@
     <div class="mapLocations shadow flex-col">
         <div class="shadow-title map-title">Active Land</div>
                 <div class="flex-row wrap">
-            <landSale v-for="land in inc.landOwned" :land="land" />
+            <land-card v-for="land in inc.landOwned" :land="land" />
         </div>
     </div>
 
     <div class="locationWrapper shadow flex-col">
         <div class="shadow-title map-title"> Land For Sale </div>
             <div class="flex-row wrap">
-            <landSale v-for="land in inc.landSale" :land="land" />
+            <land-card v-for="land in inc.landSale" :land="land" />
         </div>
     </div>
 
@@ -19,7 +19,7 @@
 </div>
 </template>
 <script>
-import landSale from '../vue/landSale.vue';
+import landCard from '../vue/landCard.vue';
 import * as helper from "../js/helper.js";
 export default{
     data () {
@@ -28,7 +28,7 @@ export default{
         }
     },
     components: {
-        'landSale': landSale
+        'land-card': landCard
     }
 }
 </script>
